+++ Authoring.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a308dac5fe2ccf64c92e684707cf654b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
