+++ Data Structures/LineStyle.cs
