+++ Obsolete/BillboardCameraProxy.cs
