+++ Obsolete/BillboardCameraProxy.cs.meta
