+++ Obsolete/LineSegmentProxy.cs
