+++ Obsolete/LineSegmentProxy.cs.meta
