+++ Obsolete/LineStyleProxy.cs
