+++ Systems/LineSegmentRegisterSystem.cs
